+++ src/components/day-page/DayPage.tsx
@@ -4,7 +4,7 @@ import MealCards from "../../features/meal-card/MealCards";
 import SavedMealCardsOffcanvas from "../../features/saved-meal/SavedMealCardsOffcanvas";
 import TabbedSummary from "../../features/summary/TabbedSummary";
 import CaloriesExceedAlertText from "../../features/warning/CaloriesExceedAlertText";
-import WarningAlert from "../../features/warning/WarningAlert";
+import WarningAlert from "../../features/warning/CaloriesExceedWarningAlert";
 import { Footer } from "./Footer";
 import { Header } from "./Header";
